tela de produto e subcategorias

+++ Controle-estoque/controle_estoque (1)/controle_estoque/src/main/java/fernandoschimidt/controle_estoque/Repository/FornecedorRepository.java
@@ -3,5 +3,5 @@ package fernandoschimidt.controle_estoque.Repository;
 import fernandoschimidt.controle_estoque.Entity.Fornecedor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FornecedorRepository extends JpaRepository<Fornecedor, Integer> {
+public interface FornecedorRepository extends JpaRepository<Fornecedor, Long> {
 }
