+++ Controle-estoque/controle_estoque (1)/controle_estoque/src/main/java/fernandoschimidt/controle_estoque/Repository/SubCategoriaRepository.java
@@ -3,5 +3,5 @@ package fernandoschimidt.controle_estoque.Repository;
 import fernandoschimidt.controle_estoque.Entity.SubCategoria;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Integer> {
+public interface SubCategoriaRepository extends JpaRepository<SubCategoria, Long> {
 }
